timer no longer runs all the time
commas
