+++ autobooks/src/configureStore.ts
@@ -1,8 +1,8 @@
 import {applyMiddleware, createStore} from "redux";
 import createSagaMiddleware from 'redux-saga'
-import sagas from './redux-store/sagas'
-import {todos} from "./App";
+import sagas from './redux-store/sagas';
 import {TimeEntry} from "./autoBooksTable/autoBooksTable";
+import {reducer} from "./redux-store/reducer";
 
 const sagaMiddleware = createSagaMiddleware();
 
@@ -20,8 +20,8 @@ const configureStore = () => {
     // let initialState = {
     //     newClockInInstances: []
     // };
-    // const store = createStore(todos, initialState, compose(applyMiddleware(sagaMiddleware)));
-    const store = createStore(todos, initialState(), applyMiddleware(sagaMiddleware));
+    // const store = createStore(reducer, initialState, compose(applyMiddleware(sagaMiddleware)));
+    const store = createStore(reducer, initialState(), applyMiddleware(sagaMiddleware));
 
     sagaMiddleware.run(sagas);
